+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 All notable changes to this project will be documented in this file.
 
 ## [UNRELEASED]
+### added
+- full CRUD support in CDO
 ### update
 - Make CDO_Record & CDO_Table track full changes